Moved the complex log joiner to it's own file.
